backend/requirements.txt
FastAPI
sqlalchemy
python-dotenv